+++ api/src/pinterest/pinterest.service.ts
@@ -58,9 +58,10 @@ interface PinterestTokenResponse {
 export class PinterestService implements SocialPlatformService {
   appCredentials: SocialProviderAppCredentials;
 
-  private readonly logger = new AppLogger(PinterestService.name);
-
-  constructor(private readonly supabaseService: SupabaseService) {}
+  constructor(
+    private readonly supabaseService: SupabaseService,
+    private readonly logger: AppLogger,
+  ) {}
 
   async initService(projectId: string): Promise<void> {
     const { data: appCredentials, error: appCredentialsError } =
