+++ api/src/telemetry/telemetry.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { ConfigModule } from '@nestjs/config';
+
+import { TelemetryService } from './telemetry.service';
+
+@Module({
+  imports: [ConfigModule],
+  providers: [TelemetryService],
+})
+export class TelemetryModule {}
